wrapped rendering function into npm package ibr-sdk (#9)
* wrapped rendering function into npm package ibr-sdk

* updated version number; included documentation for new functions
