+++ ibr/html/index.html
@@ -6,13 +6,14 @@
     <style>
 		body { margin: 0; }
 		canvas { display: block; }
-
     </style>
 </head>
 <body>
 <script src="../node_modules/pbf/dist/pbf.js"></script>
 <script src="../temp/ibr_pb_browser.js"></script>
 <script src="../binToJs.js"></script>
+<script src="../node_modules/three/build/three.js"></script>
+<script src="../node_modules/ibr-sdk/build/ibr-sdk.js"></script>
 <input type="file" id="fileForUpload" onchange='onChooseFile(event, onFileLoad.bind(this))'/>
 </body>
 </html>
