--- ibr/html/visualization.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>My first three.js app</title>
-    <style>
-		body { margin: 0; }
-		canvas { display: block; }
-
-    </style>
-</head>
-<body>
-<button onclick="back()">Back</button>
-<script src="../node_modules/three/build/three.js"></script>
-<script src="../visualization.js"></script>
-
-</body>
-</html>
