+++ ibr/ibr-sdk/src/IBR-SDK.js
@@ -1,0 +1,1 @@
+export { renderLayer } from './renderer.js';
